Urdu-OCR/requirements.txt
pytesseract
Pillow